combine definitions & remove macros
