--- src/cuckoo.cpp
@@ -1,4 +1,0 @@
-#include "cuckoo.h"
-
- std::array<uint64_t, 8192> keys{};
- std::array<Move, 8192> cuckooMoves{};
