+++ src/threads.h
@@ -10,9 +10,28 @@ enum state {
     Search,
 };
 // global vector of search threads
-extern std::vector<std::thread> threads;
+inline std::vector<std::thread> threads;
 // global vector of thread_datas
-extern std::vector<ThreadData> threads_data;
+inline std::vector<ThreadData> threads_data;
 
-[[nodiscard]] uint64_t GetTotalNodes();
-void StopHelperThreads();
+[[nodiscard]] inline uint64_t GetTotalNodes() {
+    uint64_t nodes = 0ULL;
+    for (const auto& td : threads_data) {
+        nodes += td.info.nodes;
+    }
+    return nodes;
+}
+
+inline void StopHelperThreads() {
+    // Stop helper threads
+    for (auto& td : threads_data) {
+        td.info.stopped = true;
+    }
+
+    for (auto& th : threads) {
+        if (th.joinable())
+            th.join();
+    }
+
+    threads.clear();
+}
