--- src/tune.cpp
@@ -1,21 +1,0 @@
-#include "tune.h"
-
-const int &addTune(std::string name, int defaultValue, int curr_value, int min_value, int max_value, float C_end, float R_end)
-{
-    tunable_param &param = tunableParams()[name];
-    param = tunable_param{name, defaultValue, curr_value, min_value, max_value, C_end, R_end};
-    return param.currValue;
-}
-
-bool updateTuneVariable(std::string tune_variable_name, int value)
-{
-    auto iter = tunableParams().find(tune_variable_name);
-
-    if (iter == tunableParams().end())
-    {
-        return false;
-    }
-
-    iter->second.currValue = value;
-    return true;
-}
