+++ src/tune.h
@@ -57,8 +57,25 @@ inline std::unordered_map<std::string, tunable_param> &tunableParams()
     return tunableParams;
 }
 // Actual functions to init and update variables
-const int &addTune(std::string name, int defaultValue, int curr_value, int min_value, int max_value, float C_end, float R_end);
+inline const int &addTune(std::string name, int defaultValue, int curr_value, int min_value, int max_value, float C_end, float R_end)
+{
+    tunable_param &param = tunableParams()[name];
+    param = tunable_param{name, defaultValue, curr_value, min_value, max_value, C_end, R_end};
+    return param.currValue;
+}
+
 // Handles the update of a variable being tuned
-bool updateTuneVariable(std::string tune_variable_name, int value);
+inline bool updateTuneVariable(std::string tune_variable_name, int value)
+{
+    auto iter = tunableParams().find(tune_variable_name);
+
+    if (iter == tunableParams().end())
+    {
+        return false;
+    }
+
+    iter->second.currValue = value;
+    return true;
+}
 
 TUNE_PARAM(nmpDepth, 3, 1, 5, 1, 0.002)
